+++ src/app/riesgos/scenarios/chile/tsDeus.ts
@@ -163,7 +163,7 @@ const tsTransitionProps: VectorLayerProperties = {
                 if (total > 0) {
                     [r, g, b] = yellowBlueRange(0, 1, weightedChange);
                 } else {
-                    r = g = b = 0;
+                    r = b = g = 160;
                 }
 
                 return new olStyle({
@@ -189,7 +189,7 @@ const tsTransitionProps: VectorLayerProperties = {
                             [5.627918243408203, 50.963075942052164]]]
                     }
                 },
-                text: `{{ SmallDamageChange }}`,
+                text: `SmallDamageChange`,
             }, {
                 feature: {
                     "type": "Feature",
@@ -313,7 +313,7 @@ const tsUpdatedExposureProps: VectorLayerProperties = {
                 let g: number;
                 let b: number;
                 if (total === 0) {
-                    r = b = g = 0;
+                    r = b = g = 160;
                 } else {
                     [r, g, b] = greenRedRange(0, 0.6, dr);
                 }
@@ -468,7 +468,7 @@ const tsUpdatedExposureProps: VectorLayerProperties = {
 
                 const legend = `<ul><li><b>D0:</b> {{No_damage}}</li><li><b>D1:</b> {{Minor_damage}}</li><li><b>D2:</b> {{Moderate_damage}}</li><li><b>D3:</b> {{Major_damage}}</li><li><b>D4:</b> {{ Complete_damage }}</li><li><b>D5:</b> {{ Collapsed }}</li><li><b>D6:</b> {{ Washed_away }}</li></ul>`;
 
-                return `<h4 style="color: var(--clr-p1-color, #666666);">Tsunami: {{ damage_classification }}</h4>${anchor.innerHTML}<br/>${legend}{{StatesNotComparable}}`;
+                return `<h4 style="color: var(--clr-p1-color, #666666);">Tsunami: {{ damage_classification }}</h4>${anchor.innerHTML} ${legend}{{StatesNotComparable}}`;
             },
             summary: (value: FeatureCollection | FeatureCollection[]) => {
                 let features;