+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Bug Fixes
 - **New version of Plotly.js:** Fixes bug where graph axes were not labeled.
 - **Flood damage:** Products were not highlighted when associated process is selected.
+- **Focus on selected earthquake:** Preliminarily selected earthquake is no longer a mappable-product. Only the finally selected one is. This way the right output product is highlighted in the layer-selection when a user focuses on the eq-selection step.
 
 ### Features
 - **Removed @dlr-eoc/services-util-storage:** Better build time, less optimization bailouts due to module formats CommonJS or AMD.
@@ -14,6 +15,7 @@
 - **Using same data-source for all tsunami-damage-layers:** Saves a lot of memory
 - **Added vector-tile base-map:** Allows individually styling background layer. Adjusts to dark-mode.
 - **Spectral acceleration data:** Now also displayed as another output of the ground-motion-service.
+- **Colors:** updated color scales to be less saturated.
 
 ### Breaking Changes
 - **xyz:**